Dockerfile is modified to prevent libprotobuf-mutator tests from being executed during the Docker image creation.
It was discovered that starting from a specific version of Clang,
the deprecated-copy warning occurs when building googletest,
which is used for testing libprotobuf-mutator.
This warning is treated as an error due to the build flags,
causing the final docker build to fail.

Since this process is not necessary and the issue is a warning
rather than an actual error, we have made a patch to eliminate
this unnecessary step.